refactoring logging integration test
